requirement.txt
flask
pandas
numpy
tensorflow
matplotlib
scikit-learn
openpyxl
pillow